refactor product category service
