+++ libs/domain/product/src/models/category-qualifier.ts
@@ -1,5 +1,7 @@
 export interface ProductCategoryQualifier {
   id?: string;
+  /** limit result to direct children categories of the parent */
   parent?: string;
+  /** exclude category or cateories from the result */
   exclude?: string | string[];
 }
